* all: Removed scheduler functionality from Core.c. Added Scheduler.c/h
